Refactor to move async version next to sync
Removes the httpio_async package and moves the code into the
httpio.asyncio module.
Makes integration with our build system easier.

sem-ver: api-break
